05_embeddings/requirements.txt
langchain==0.0.228 
llama_index==0.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2